Update billing code to be opt-in
We now return claimed_time with the query behind payday, and filter on
that when computing the total to charge a tipper's credit card and how
to distribute their gifts. This addresses #28. We also now have that
info (claimed_time) available on the tipper's profile page. We should
indicate on their which of their tips are to people who have claimed
their account and which are not.