Change from "gift economy" to "meta gift economy"
I had a conversation on twitter with anatsuno, jinjurly, krytella, and
teganmul about whether or not Gititp is a gift economy. It seems that
Gittip is more of "a tool for gift economies" than a direct gift economy
itself. I'm trying out the phrase "meta gift economy" to capture this
nuance, and I've added a section to the About page to explain it (as
well as reworking a few other parts of the page in light of same). #81