Now we're a "micro-giving platform"
Slept on the conversation re: whether Gittip is a gift economy:

  http://storify.com/jinjurly/tips-and-economies-and-expectations-oh-my

I care more about the reality than the label. I'm content to let others
make the connection with gift economics when the time is right.
Certainly there are differences from the classic gift economies in what
Gittip is doing.

I also would prefer not to make strong statements about the vision that
Gittip is pursuing ("reversing the polarity of the economy," etc.)
because I want Gittip to have mass appeal, and people are going to all
have their own agendas and ideas. Gittip should be an ideologically
neutral system that as many people as possible can get excited about and
come together over. Of course people are free to express their view of
and desires for Gittip, but there shouldn't be an "official" ideology.

In this light I removed the Gift Economy section of the about page, and
changed the tag from "Gittip is a gift economy" to "Gittip is a
micro-giving platform."