Derp, here's the meat of that last commit (#172)
This also drops back to not listing the card type of the card on file,
and showing ************1234 as before (doing so actually flushed out a
bug in gittip.billing.BalancedAccount). The reason for this is that if
we're going to show the card type I want to capitalize it properly and
use the proper article (a or an), and that's deeper than I want to go at
this point.