Take out broken error handling on cc/bank pages
I made an attempt a while back to provide friendlier error messages when
the credit card or bank account form failed. However, the way it was
implemented meant that such errors would not get logged to Sentry and
we're unlikely to find out about them: Bad! Until we can do both
(friendly messages *and* logging to sentry), I'm dropping back to just
logging to Sentry.