Make initial signin friendlier; #634
This varies the message for people with no giving or receiving based on
how long they've been on the site. For the first minute, we say "just
joined Gittip!" then for a week we say "joined recently," then after
that we go with "is lurking." It was too abrasive to say "is lurking" as
the first thing people see when they join the site.