Fix bug with updating amounts; #635
This barely fixes the immediate problem. We really would need to
refactor to implement this properly, and possibly bring on a client-side
framework (Backbone, Ember, Angular, etc.). For example, we don't handle
the case when a tip change would make receiving greater than receiving,
so that we want to flip the display of those values in the box/card. Or
if this is the first tip received, we don't add that in. That kind of
thing.