Implement UI for recording manual exchanges; #53
Sometimes we do manual payouts (maybe payins someday?) and we need a way
to record those in the database so we don't have to keep doing them
directly in SQL(!). This adds a form to the history page for site admins
to use. It extends the exchanges schema to keep track of who made manual
adjustments and why.