Take GA off widget.html; #824
For the widget it really slows us down to have tracking code on there.
Removing all js tracking is #826 but for now let's at least only use
one. I'm going for Gauges. The UI is much nicer and this will give us a
chance to kick the tires a bit. The widget traffic in GA has always been
kind of annoying.