Enough fucking around. We need this. #449
I've got people pressuring me to turn this into a for-profit business so
we can (take VC?) and pay people salaries and make it grow. And they're
right. Gittip needs to be funded in order to grow. No two ways about it.
But I'll be damned if I'm taking VC or profit. Gittip is funded on
Gittip.