Fix bug in community giver/receiver amounts
By using _wrap as a query_cache post-processing function we were ending
up with a single Participant object being used for both giver and
receiver lists when the same user was both a giver and a receiver. The
problem is that the amount giving is scabbed onto that object the the
amount received was shown in both givers and receivers lists. This goes
back to _wrap as a generator around those lists.