Clean up error handling; #983
In the original pull request (#933), when Bountysource gave us an error
in the querystring to ./associate, we displayed that error message on
the user's profile page. That's not the worst idea in theory, but in
practice we don't have good UI infrastructure in place for that. On the
pattern of lock-fail.html in the other account elsewhere
implementations, this displays the Bountysource error message in a
dedicated ./problem.html page.