Rationalize make targets for db; #1078
We used to have `data`, which created the gittip-test database, and
`fake_data`, which populated the database configured in local.env. Now we
have:

 - `db` - creates the database configured in local.env
 - `data` - populates the database configured in local.env
 - `test-db` - creates the database configured in tests/env

We don't need a `test-data` target because for tests we create data
granularly for each test case. The `data` target is to populate a local
instance with fake data so you have a filled-out app to browse through
the web on localhost:8537.