Take out separate brake; #1090
We really had two brakes, the 1.5x rule and the 10/count(members) rule.
We should only have one rule if we can to keep this simple and
intuitive. The 1.5x rule is enough of a constraint to prevent an
untrusted person from taking all the money in a single week.