Further automate building docs from docstrings
Sphinx has these automodule directives for building docs out of
docstrings, but the way that works is that you have to manually tell
Sphinx which modules to document in this way. There's an `sphinx-apidoc`
cli for generating the necessary *.rst files for you, but it doesn't
suit our needs:

 - We want one doc file per module.
 - We don't want "Module" or "Package" in the titles.
 - We want to use directories to organize doc files instead of dotted
   names.

This commit adds a build-rst.py script that generates *.rst files for
the gittip library the way we want.