Add a backup script
This adds a little more automation to our weekly payday routine. The
first thing I do is take a backup of the database, which I store
locally. This automates that process. I factored out a script to convert
the DATABASE_URL envvar into PG* envvars for use by pg_dump (backup.sh)
and psql (recreate-schema.sql). It now echoes the envvars rather than
exporting as before when it was part of recreate-schema.sql.