Fix regressions in community chooser on /for/
We used to have the community drop-down on both /for/ and the profile
page. We took it off the profile page in #1087. Now that we are
including the profile.js on every page, it was clobbering the drop-down
on /for/. I've moved the document.ready in profile.js (are there
others?) into an init function that's only fired on profile edit. I also
moved the rendering of the community listing on profiles server-side.