Clean up communities on profile some more
There was a typo, nother instead of nothers, that was doubly masked.
First, if you had a team then nothers was left over from that loop.
You'd expect the numbers for the communities to all be the number for
the last team in that case, but we still had the js that was setting the
numbers client-side, so if you had a team then everything looked fine.
It was only if you didn't have a team that you saw the nother NameError.