Fix test regression introduced by #1457 fix
This one is a little convoluted. We introduced the cent sign unicode
character into our mark-up at some point (#1070?), and that should've
been in the output of the self.get call in this case but wasn't because
of the #1457 bug. The test still passed because it's not a test for the
cent sign.