Force 127.0.0.1:8537 to avoid localhost confusion
It's more normal to use localhost than 127.0.0.1, but Twitter won't let
us use localhost as the location or callback. We can use 127.0.0.1,
however, but since people are used to typing localhost let's redirect
them to 127.0.0.1 so they don't run into issues when they try to login
with our default dev oauth apps and land on a different host.