Rough-in of an elsewhere refactor refactor; #1369
This adds a PlatformRegistry to website, which provides access to
instances of Platform subclasses, which have a load method that takes a
username on the given platform and returns an AccountElsewhere
orm.Model. This is all barely working with one test. Surely other tests
are broken now, I haven't even checked. Gonna start building back up and
out.