Rip out _resolve
Afaict this is straight cruft. I don't see any actual usage of it in the
codebase right now. The old version of it took a platform and username
on the platform and returned a username on Gittip. That's redundant now
that we have participant objects on our account elsewhere objects.

The one thing I thought it might be for is redirecting account elsewhere
URLs to Gittip profiles in the case of claimed accounts. I implemented
that here.