This adds a participants.paypal_email field
The purpose of this field is to record a PayPal email address for people
whom we're paying out via MassPay. Since MassPay isn't intended as a
long-term option, I'm proposing that we don't need a UI for this field.
I'll update it manually via psql as needed. Generally I hear from people
on support@gittip.com. I confirm the email address manually by asking
for the first eight digits of their API key.