faq: reword Coinbase sentence to avoid confusions
"Using Coinbase" is an implementation detail that could lead
people to think they need a Coinbase account for getting the
payout, which is not true because a bitcoin address is the
only requirement (Coinbase is just the USD->BTC conversion
service), as pointed out in #310.

So we reword the sentence to avoid this confusing but still
mention Coinbase as an informative detail about what rates
and fees apply.