Flip flag around; #1667
When the flags go right then they overlap any chart to the right. We had
a fix in #1677, but it was kinda a hacky solution and it didn't adapt
well to changing chart widths and I also clobbered it when merging
master out to the #1667 branch. So here's a different fix for the same
problem: flip the flags.