allow charts to have sub-pixel widths
By removing this restriction, our bar graphs can now stretch to fill
all of the available space instead of wasting the space that wasn't
evenly divisible into whole pixels for each bar.

Not all browsers support sub-pixel rendering, so the code reverts back
to the previous behavior instead of rendering an ugly graph to the user.