Handle the Balanced API more saferly during payday
This makes us more robust in the face of failures communicating with the
Balanced API. Now we've got a blanket try/except around the parts where
we talk to Balanced, so we will never totally crash because of a
mishandling of some sort. The ach_credit side is not as well-factored
as the charge_on_balanced side, and is not tested as a result. This
commit nonetheless attempts to implement the same error handling on both
sides.