Take out probably misguided check for merchants
At https://github.com/gittip/www.gittip.com/pull/2036/files#r9757101
@mjallday suggests that we don't need to check that merchants are
underwritten, as that constraint now only applies to those receiving >
$100,000/yr, and we don't have anyone there yet. Moreover, I believe
this condition is backwards, and would result in no-one getting paid!
Without any kind of condition here, I suspect that we'll fill our logs
with NoResultError()s for everyone who is due a payout and has a CC on
file with Balanced but not a bank account. Better that than not doing
any payouts. :-)