Move account row to a macro
The "Connected Accounts" listing includes accounts elsewhere along with
some kludged together systems (Bitcoin, Balanced). As part of #2053 we
want to group the items under Connected Accounts, but the groups don't
fall out along the lines of the accounts elsewhere/kludge split. Moving
this bit of markup to a macro gives us the flexibility we need to knit
true and kludges accounts elsewhere into the UI we want.