Here's yet another attempt to placate PayPal
It turns out there's another twist to the MassPay penny saga: in some
situations (when amounts are small enough?), the fee is rounded to what
we wanted it to be anyway. See #2296 for a description.

This commit takes a whole 'nother approach to fee computation. Instead
of computing / 1.02 of the target and then fudging it from there, we
take a much simpler approach. We start at a penny under the target and
work down from there a penny at a time until we find a value that works.
Less efficient but safer.