Add a (formerly Gittip) reference to about page
Currently there is a lot of 3rd party places that still reference gittip, and generally it will take a while for the name change to become common knowledge. A lot of articles, press, and links that have already been written, will maintain the old Gittip name.

As such, it is important that Gratipay makes reference that it was once Gittip to maintain the trust. As say someone reads an article on gittip, or goes to donate via a gittip link, now they are taken to Gratipay, wait what, is this a scam? Who the hell is Gratipay, I wanted Gittip.

Hopefully this minor change is a first step to solving that confusion, but more steps are probably needed.