Update Giving page
I dropped back to more standard table styling. I went ahead and dropped
the inline tip setting UI. It was really ugly and cumbersome. Instead,
users will have to now click through to the user in question to change
their tip in the sidebar on the recipient's profile. I did add the
ctime/mtime. We can make these tables sortable in the future (along with
the members table for teams).