Standardize on no _underscore.scss
There's a convention in the Sass community to prefix partials with
underscore. It's actually more than a convention, because:

`scss --watch dirname`

will produce a css file for every scss file it finds *unless* the
filename starts with underscore. Because we're not using this method of
compiling scss (we programmatically compile our scss from a simplate),
we don't need to worry about this. Let's skip the underscores.