Move tip distribution computation to the server
I'm porting the tip distribution charts to the same library we're using
for the main charts, so that I can fold the charts page into the stats
page. This is part of an overall rationalization of the information
architecture under /about/.

As part of this I'm switching to bins that make sense to humans, based
on 1s, 2s, and 5s.

The next step is to generalize gratipay/charts.js to work with a
non-temporal X axis.