Here's a global notice about site disruption
I tried implementing this using our js notification system, but that
doesn't support HTML that I can tell. I wanted to include links. The
content area is the best place I could quickly find to put this, but
that's not available on the unauthenticated homepage, so as a quick hack
this PR redirects / to /about/ for anon.