Update new status field when balance goes to 0
There are going to be two ways to resolve a Gratipay 1.0 balance, either
we release it, or we refund it. We release it the first payday after
someone becomes a team owner, or (now) when we set `pending-payout`. At
this point I don't think we'll need a `pending-refund` value. I've
changed `{un,}released` to `{un,}resolved` to cover both cases, release
and refund. I implemented the status update in Postgres to cover both
payday and masspay.