rejected -> closed
"Rejecting" teams is too harsh a term for the situation we want to
optimize for, wherein we reach a decision collaboratively together with
team owners about whether they and we are a good fit for each other. I
think we should use the term "closed" instead. That leaves ambiguous the
circumstances surrounding the closing, whereas "rejected" only covers
the harshest case, wherein Gratipay acts despite the owner's intention.
A "closed" account could be closed by the owner's decision, by
Gratipay's, or by both agreeing together.