Remove update_cached_amounts
This is no longer necessary, because we're not allowing ~users to fund
payments entirely out of their taking. They have to have a credit card
attached. If they do, we'll only charge them if their takes in the same
week don't cover it.

However, we now need to update `giving` every time we charge a card,
since we'll no longer pick up card failures at the end of payday.