Decouple status update from deployment
Setting the status_of_1_0_payout field depends on the numbers_1_0 table,
which is expensive to create. Let's keep the schema changes in
branch.sql but move the status update to a separate script that we can
run once deployment is finished. This'll require a second PR to drop the
status_of_1_0_balance column/function/trigger once we're done.