Conditionally fall back to pure-Python ijson
At Travis and in production we still want to require YAJL, for
performance, but we don't want to require that in local development.
Note that we still skip the relevant tests if marky-markdown is not
installed, so as to avoid introducing a Node dependency to local
production. Unlike with ijson, which would silently degrade if local
conditions were applied to production, a missing marky-markdown will
hard-fail in production.