Update sandbox Braintree credentials
For whatever reason they are different now when I browse to find them in
the sandbox admin. This also folds back to one set of creds for both dev
and test. We split test in 6d8b0a14086348c260d1690739295f0654a8f6c1 but
I'm not sure why (convenience at the time?).